Initial commit:
An oversetAndAMRFvMesh object created which is a copy of static dyanamicFvMesh.
Then I added two private members to it:

    // overset mesh member
    dynamicOversetFvMesh meshOverset_;

    // dynamic refine member
    dynamicRefineFvMesh meshDynamicRefine_;

The plan is to call
meshOverset_.update() && meshDynamicRefine_.update()

in this object's update method.

--------------
But when I am doing that it is complaining that U and p are not found in the object registry.